airflow_requirements.txt
apache-airflow
apache-airflow-providers-docker
airflow.operators.bash_operator
airflow.operators.python
requests